+++ backend/techpick-api/src/main/java/kernel360/techpick/api/api/pick/controller/PickApiController.java
@@ -16,11 +16,11 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import kernel360.techpick.api.api.pick.dto.PickApiMapper;
+import kernel360.techpick.api.api.pick.dto.PickApiRequest;
+import kernel360.techpick.api.api.pick.dto.PickApiResponse;
+import kernel360.techpick.api.domain.pick.service.PickService;
 import kernel360.techpick.core.annotation.LoginUserId;
-import kernel360.techpick.feature.api.pick.dto.PickApiMapper;
-import kernel360.techpick.feature.api.pick.dto.PickApiRequest;
-import kernel360.techpick.feature.api.pick.dto.PickApiResponse;
-import kernel360.techpick.feature.domain.pick.service.PickService;
 import lombok.RequiredArgsConstructor;
 
 @RestController
