--- backend/techpick-api/src/main/java/kernel360/techpick/api/application/PickFacade.java
@@ -1,4 +1,0 @@
-package kernel360.techpick.api.application;
-
-public class PickFacade {
-}
