+++ backend/techpick-api/src/main/java/kernel360/techpick/api/domain/pick/dto/PickCommand.java
@@ -9,16 +9,16 @@ public class PickCommand {
 	public record Read(Long userId, Long pickId) {
 	}
 
-	public record Create(Long userId, String title, String memo, List<Long> tagOrder, Long parentFolderId,
+	public record Create(Long userId, String title, String memo, List<Long> tagOrderList, Long parentFolderId,
 						 LinkInfo linkInfo) {
 	}
 
 	public record Update(Long userId, Long pickId, String title, String memo, List<Long> tagIdList) {
 	}
 
-	public record Move(Long userId, Long pickId, Long parentFolderId, int orderIdx) {
+	public record Move(Long userId, List<Long> pickIdList, Long destinationFolderId, int orderIdx) {
 	}
 
-	public record Delete(Long userId, Long pickId) {
+	public record Delete(Long userId, List<Long> pickIdList) {
 	}
 }
