+++ backend/techpick-api/src/main/java/kernel360/techpick/api/domain/pick/dto/PickResult.java
@@ -1,5 +1,6 @@
 package kernel360.techpick.api.domain.pick.dto;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import kernel360.techpick.api.domain.link.dto.LinkInfo;
@@ -9,6 +10,9 @@ public record PickResult(
 	String title,
 	String memo,
 	LinkInfo linkInfo,
-	List<Long> tagOrder
+	Long parentFolderId,
+	List<Long> tagOrderList,
+	LocalDateTime createdAt,
+	LocalDateTime updatedAt
 ) {
 }
