+++ backend/techpick-api/src/main/java/kernel360/techpick/api/domain/pick/service/PickService.java
@@ -1,5 +1,7 @@
 package kernel360.techpick.api.domain.pick.service;
 
+import java.util.List;
+
 import kernel360.techpick.api.domain.pick.dto.PickCommand;
 import kernel360.techpick.api.domain.pick.dto.PickResult;
 
@@ -7,11 +9,13 @@ public interface PickService {
 
 	PickResult getPick(PickCommand.Read command);
 
+	PickResult getPickUrl(Long userId, String url);
+
 	PickResult saveNewPick(PickCommand.Create command);
 
 	PickResult updatePick(PickCommand.Update command);
 
-	PickResult movePick(PickCommand.Move command);
+	List<Long> movePick(PickCommand.Move command);
 
 	void deletePick(PickCommand.Delete command);
 }
