+++ backend/techpick-api/src/main/java/kernel360/techpick/api/infrastructure/link/LinkAdaptor.java
@@ -1,0 +1,11 @@
+package kernel360.techpick.api.infrastructure.link;
+
+import kernel360.techpick.api.domain.link.dto.LinkInfo;
+import kernel360.techpick.core.model.link.Link;
+
+public interface LinkAdaptor {
+
+	Link getLink(String url);
+
+	Link saveLink(LinkInfo linkInfo);
+}
