+++ backend/techpick-api/src/main/java/kernel360/techpick/api/infrastructure/link/LinkAdaptorImpl.java
@@ -1,18 +1,15 @@
-package kernel360.techpick.feature.infrastructure.link;
+package kernel360.techpick.api.infrastructure.link;
 
 import java.util.Optional;
 
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.zaxxer.hikari.util.IsolationLevel;
-
+import kernel360.techpick.api.domain.link.dto.LinkInfo;
+import kernel360.techpick.api.domain.link.dto.LinkMapper;
+import kernel360.techpick.api.domain.link.exception.ApiLinkException;
 import kernel360.techpick.core.model.link.Link;
 import kernel360.techpick.core.model.link.LinkRepository;
-import kernel360.techpick.feature.domain.link.dto.LinkInfo;
-import kernel360.techpick.feature.domain.link.dto.LinkMapper;
-import kernel360.techpick.feature.domain.link.exception.ApiLinkException;
 import lombok.RequiredArgsConstructor;
 
 @Component
