--- backend/techpick-api/src/main/java/kernel360/techpick/api/infrastructure/link/reader/LinkReader.java
@@ -1,8 +1,0 @@
-package kernel360.techpick.api.infrastructure.link.reader;
-
-import kernel360.techpick.core.model.link.Link;
-
-public interface LinkReader {
-
-	Link read(String url);
-}
