+++ backend/techpick-api/src/main/java/kernel360/techpick/api/infrastructure/pick/PickAdaptor.java
@@ -1,16 +1,34 @@
 package kernel360.techpick.api.infrastructure.pick;
 
-import kernel360.techpick.core.model.pick.Pick;
-import kernel360.techpick.core.model.user.User;
+import java.util.List;
+
+import kernel360.techpick.api.domain.pick.dto.PickCommand;
 import kernel360.techpick.api.domain.pick.exception.ApiPickException;
+import kernel360.techpick.core.model.pick.Pick;
+import kernel360.techpick.core.model.pick.PickTag;
+import kernel360.techpick.core.model.tag.Tag;
 
 public interface PickAdaptor {
 
-	Pick getPick(User user, Long pickId);
+	Pick getPick(Long pickId);
 
-	Pick savePick(Pick pick) throws ApiPickException;
+	Pick getPickUrl(Long userId, String url);
 
-	void deletePick(Pick pick);
+	List<Pick> getPickList(List<Long> idList);
+
+	List<PickTag> getPickTagList(Long pickId);
+
+	Pick savePick(PickCommand.Create command) throws ApiPickException;
+
+	PickTag savePickTag(Pick pick, Tag tag);
+
+	Pick updatePick(PickCommand.Update command);
+
+	void movePickToCurrentFolder(PickCommand.Move command);
+
+	void movePickToOtherFolder(PickCommand.Move command);
+
+	void deletePickList(PickCommand.Delete command);
 
 	void detachTagFromPick(Pick pick, Long tagId);
