--- backend/techpick-api/src/main/java/kernel360/techpick/api/infrastructure/user/writer/UserWriter.java
@@ -1,4 +1,0 @@
-package kernel360.techpick.api.infrastructure.user.writer;
-
-public interface UserWriter {
-}
