--- backend/techpick-api/src/main/java/kernel360/techpick/api/infrastructure/user/writer/UserWriterImpl.java
@@ -1,4 +1,0 @@
-package kernel360.techpick.api.infrastructure.user.writer;
-
-public class UserWriterImpl implements UserWriter {
-}
